refactor(api): move usecase detach target profil to bounded context target profile
